Core Mod, Concurrent Particles Patch
Framework for the core mod (so much for this being a simple tweaker...)

Patch to move particles to a ConcurrentLinkedDeque (from ArrayDeque), this will fix any issues where the logical server creates particles on the physical client.

Uses the incredibly helpful ASMHelper package, which was released under Unlicense

I setup a kind of weird flexible subclassing patch system... it'll work.
Need to work on better ClassNode transformation utilities...