Remove Village and Fortress Cleanup
If you enter a village or fortress, unload the chunks, have the world save twice, and then come back, the village and fortress will be broken.

Restarting the game or server, however, brings them back, which I don't quite understand, but it's a good thing, as it means this feature might be workable on server stopping.

For now, the tweak is removed. Mineshafts are still cleaned up though as this bug hardly matters for them.