Spawner Control World Tick
Remove the world tick entirely...
All it does is clean up obsolete spawners and add the smoke effect.
In return, it causes a high amount of lag, especially in packs with lots of mob spawners.

Obsolete spawners should be rare or nonexistent, only caused by changing config for the same world, so this should be fine.